Prevent the tooltip from leaking out of the chart
This was a tricky bug: we throttle the tooltip for performance reasons but because of this, pending executions could still run after the tooltip was hidden because the cursor was out of the chart. By calling the cancel method of the throttled function, we can ensure this doesn't happen anymore.