Fix issues with the line chart displaying really high charts
The y2 property was defined for no reason on the mark and set to 0, which would force the chart to use a scale that contains 0.

Because of that, if the data would be "centered" far away from 0 or with a great precision around a point that isn't 0, the chart could get really high (> 30M px).