Remove the prop to embed a widget in the WidgetEditor component
A widget can't be embedded until it's registered in the API. Also, it's been pointed out by David G. that it could be confusing to have an embed option in the editor because the user might make changes that haven't been saved yet. Finally, the code was somewhat wrong and the Share/Embed option was actually never shown.