Fix an issue where bars could overlap the x axis
It actually seems to be a bug from Vega: the null values are computed as "0" but don't count in the calculation of the y scale. That causes the bar to overlap the x axis when the the option "zero" is set to false in the vega configuration. After talking with Alicia, the vertical scale should always start with 0, so the problem is easily fixed.