Prevent the editor from breaking when restoring an outdated widget
If the editor would be restored with a widget whose configuration contains columns that don't exist anymore, it would try to render it before checking if it's correct leading to errors.

A flag has been added in the state to monitor the "initialization" of the editor before rendering any visualization.