Get custom Renderer usage working
There is specific styling that is required for the Renderer to work.
This doesn't determine exactly what that styling is, but it does apply
existing CSS classes that are sufficient (specifically, by nesting
`.widget-container` inside `.c-widget-block`). If we need to re-create
these rules later then we can use this as a reference.