Updated query for `minerals-energy` on country page
I noticed many counties are displaying 0 for this indicator because of small values. The new query should return the rounded value if it is >0, `<1` if the value is between 0 and 1, and `0` if the value is 0. Since this query returns a string, I am wondering if the percent sign should be included in the `unit` parameter or in the query itself.